backend/requirements.txt
# ========================================================
# Backend Dependencies - RAG QABot Auth/API Service
# ========================================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database (PostgreSQL + SQLAlchemy 2.0) ---
sqlalchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.10

# --- Redis (Upstash compatible) ---
redis==5.1.1

# --- Auth & Security ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# --- Config & Validation ---
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# --- HTTP Client (for internal calls) ---
httpx==0.27.2
